src/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2